requirements.txt
flask==3.0.3
flask-restx==1.3.0
pillow==10.4.0
Flask-Testing==0.8.1
pytest==8.3.2
pytest-cov==5.0.0
mock==5.1.0